Refactor to use mixins
