+++ src/mixins/index.ts
@@ -1,0 +1,2 @@
+export * from './AlertMixin';
+export * from './ValidationMixin';
