+++ src/types/ValidationError.ts
@@ -1,0 +1,4 @@
+export default interface ValidationError {
+  code: string;
+  message?: string;
+}
