+++ src/views/Fridge.vue
@@ -56,7 +56,7 @@
 
 <script lang="ts">
 import router from '@/router';
-import { AlertMixin } from '@/components/AlertMixin';
+import { AlertMixin } from '@/mixins/AlertMixin';
 import { Component, Mixins } from 'vue-property-decorator';
 import { CurrentFamily } from '@/types';
 import Firestore from '@/utils/Firestore';
