+++ src/views/ShoppingList.vue
@@ -40,7 +40,7 @@
 
 <script lang="ts">
 import router from '@/router';
-import { AlertMixin } from '@/components/AlertMixin';
+import { AlertMixin } from '@/mixins/AlertMixin';
 import { Component, Mixins } from 'vue-property-decorator';
 import { CurrentFamily } from '@/types';
 import Firestore from '@/utils/Firestore';
